refactor(Listbox): use `compare` and `valueComparator` functions from `@/shared`
