+++ packages/core/src/Listbox/ListboxVirtualizer.vue
@@ -21,10 +21,10 @@ import { refAutoReset } from '@vueuse/shared'
 import { cloneVNode, computed, Fragment, useSlots } from 'vue'
 import { useCollection } from '@/Collection'
 import { MAP_KEY_TO_FOCUS_INTENT } from '@/RovingFocus/utils'
-import { findValuesBetween, getActiveElement } from '@/shared'
+import { compare, findValuesBetween, getActiveElement } from '@/shared'
 import { getNextMatch } from '@/shared/useTypeahead'
 import { injectListboxRootContext } from './ListboxRoot.vue'
-import { compare, queryCheckedElement } from './utils'
+import { queryCheckedElement } from './utils'
 
 const props = defineProps<ListboxVirtualizerProps<T>>()
